+++ src/test/vlocityDatapackService.test.ts
@@ -17,17 +17,17 @@ describe('vlocityDatapackService', () => {
     describe('#resolveProjectPathFor (private)', () => { 
         it("should always return absolute path", function() { 
             let vds = [
-                new vlocityDatapackService({ projectPath: '.\\project\\vlc\\' }),
-                new vlocityDatapackService({ projectPath: '.\\' }),
-                new vlocityDatapackService({ projectPath: '.\\project\\folder\\..' }),
-                new vlocityDatapackService({ projectPath: '.\\project' })
+                new vlocityDatapackService(null, { projectPath: '.\\project\\vlc\\' }),
+                new vlocityDatapackService(null, { projectPath: '.\\' }),
+                new vlocityDatapackService(null, { projectPath: '.\\project\\folder\\..' }),
+                new vlocityDatapackService(null, { projectPath: '.\\project' })
             ];
             let uriForResolve = vscode.Uri.file('./myfile');
             // assert
             vds.forEach(v => expect(path.isAbsolute((<any>v).resolveProjectPathFor(uriForResolve))).equals(true));
         });
         it("should return project path as-is when absolute", function() { 
-            let vds = new vlocityDatapackService({ projectPath: 'x:\\project\\vlc\\' });
+            let vds = new vlocityDatapackService(null, { projectPath: 'x:\\project\\vlc\\' });
             let uriForResolve = vscode.Uri.file('./myfile');
             // assert
             expect((<any>vds).resolveProjectPathFor(uriForResolve)).equals('x:\\project\\vlc\\');
@@ -37,14 +37,14 @@ describe('vlocityDatapackService', () => {
     describe('#getDatapackManifestKey',  () => {
         it('should resolve single level datapacks', () => {
             let datapackFile = vscode.Uri.file('c:\\project\\vlc\\dptype\\dpname\\mypack_datapack.json');
-            let vds = new vlocityDatapackService({ projectPath: 'c:\\project\\vlc\\' }).getDatapackManifestKey(datapackFile);
+            let vds = new vlocityDatapackService(null, { projectPath: 'c:\\project\\vlc\\' }).getDatapackManifestKey(datapackFile);
             // assert
             expect(vds.datapackType).equals('dptype');
             expect(vds.key).equals('dpname');
         });
         it('should resolve multi level datapacks', () => {
             let datapackFile = vscode.Uri.file('c:\\project\\vlc\\dptype\\dpname1\\dpname2\\mypack_datapack.json');
-            let vds = new vlocityDatapackService({ projectPath: 'c:\\project\\vlc\\' }).getDatapackManifestKey(datapackFile);
+            let vds = new vlocityDatapackService(null, { projectPath: 'c:\\project\\vlc\\' }).getDatapackManifestKey(datapackFile);
             // assert
             expect(vds.datapackType).equals('dptype');
             expect(vds.key).equals('dpname1/dpname2');
@@ -72,7 +72,7 @@ describe('vlocityDatapackService', () => {
         });
         it('should resolve datapack header for folders/directories', (done) => {
             let datapackFolder = vscode.Uri.file('c:/datapacks/product2/test1');
-            new vlocityDatapackService().resolveDatapackHeader(datapackFolder).then(datapackHeader => {
+            new vlocityDatapackService(null, {}).resolveDatapackHeader(datapackFolder).then(datapackHeader => {
                 // assert
                 expect(datapackHeader).not.equals(undefined);
                 expect(datapackHeader.path).equals('/c:/datapacks/product2/test1/header_Datapack.json');
@@ -81,7 +81,7 @@ describe('vlocityDatapackService', () => {
         });
         it('should resolve datapack header for siblings', (done) => {
             let datapackFolder = vscode.Uri.file('c:/datapacks/product2/test1/otherfile.html');
-            new vlocityDatapackService().resolveDatapackHeader(datapackFolder).then(datapackHeader => {
+            new vlocityDatapackService(null, {}).resolveDatapackHeader(datapackFolder).then(datapackHeader => {
                 // assert
                 expect(datapackHeader).not.equals(undefined);
                 expect(datapackHeader.path).equals('/c:/datapacks/product2/test1/header_Datapack.json');
@@ -90,7 +90,7 @@ describe('vlocityDatapackService', () => {
         });
         it('should resolve datapack header when datapack header is passed', (done) => {
             let datapackFolder = vscode.Uri.file('c:/datapacks/product2/test1/header_Datapack.json');
-            new vlocityDatapackService().resolveDatapackHeader(datapackFolder).then(datapackHeader => {
+            new vlocityDatapackService(null, {}).resolveDatapackHeader(datapackFolder).then(datapackHeader => {
                 // assert
                 expect(datapackHeader).not.equals(undefined);
                 expect(datapackHeader.path).equals('/c:/datapacks/product2/test1/header_Datapack.json');
@@ -99,7 +99,7 @@ describe('vlocityDatapackService', () => {
         });
         it('should return undefined for files outside of a datapack folder', (done) => {
             let datapackFolder = vscode.Uri.file('c:/datapacks/product2/test2/no_dataoacks_here.html');
-            new vlocityDatapackService().resolveDatapackHeader(datapackFolder).then(datapackHeader => {
+            new vlocityDatapackService(null, {}).resolveDatapackHeader(datapackFolder).then(datapackHeader => {
                 // assert
                 expect(datapackHeader).equals(undefined);
                 done();
@@ -107,7 +107,7 @@ describe('vlocityDatapackService', () => {
         });
         it('should return undefined for non-existoing files', (done) => {
             let datapackFolder = vscode.Uri.file('c:/datapacks/product2/test2/no_dataoacks_here.html');
-            new vlocityDatapackService().resolveDatapackHeader(datapackFolder).then(datapackHeader => {
+            new vlocityDatapackService(null, {}).resolveDatapackHeader(datapackFolder).then(datapackHeader => {
                 // assert
                 expect(datapackHeader).equals(undefined);
                 done();